refactor:#74 기능별로 코드 분리
